Move getGossipPeers to main class
